refactor: refactor Json processing
...when reading Json instances.

ING-4046
